+++ docs/html/search/all_5.js
@@ -1,0 +1,14 @@
+var searchData=
+[
+  ['get_5fclients',['get_clients',['../classtacopie_1_1tcp__server.html#a0df81b943243ad51102c37d4944be8d7',1,'tacopie::tcp_server']]],
+  ['get_5ffd',['get_fd',['../classtacopie_1_1tcp__socket.html#a4fd367a1802ed6ae7fa0e33fc07ed255',1,'tacopie::tcp_socket']]],
+  ['get_5ffile',['get_file',['../classtacopie_1_1tacopie__error.html#a10360163c780a1bd9c95fcecca5aa6da',1,'tacopie::tacopie_error']]],
+  ['get_5fhost',['get_host',['../classtacopie_1_1tcp__client.html#ad38ab710c5eca64de2f887abc455b15d',1,'tacopie::tcp_client::get_host()'],['../classtacopie_1_1tcp__socket.html#ad294565f9a0fa52639ecfbf133eecd59',1,'tacopie::tcp_socket::get_host()']]],
+  ['get_5fio_5fservice',['get_io_service',['../classtacopie_1_1tcp__client.html#aafbf0aa37cd0472778d09fb163362314',1,'tacopie::tcp_client::get_io_service()'],['../classtacopie_1_1tcp__server.html#aace4796627b6abccccce1a541908414f',1,'tacopie::tcp_server::get_io_service()']]],
+  ['get_5fline',['get_line',['../classtacopie_1_1tacopie__error.html#a39704d2cb6f076aa47d45f53f174b257',1,'tacopie::tacopie_error']]],
+  ['get_5fport',['get_port',['../classtacopie_1_1tcp__client.html#a3b42ae2afe6d5ee5f2f16b8bd7846f37',1,'tacopie::tcp_client::get_port()'],['../classtacopie_1_1tcp__socket.html#a5276fdc687ac3c5089a05e3e2d9de4fb',1,'tacopie::tcp_socket::get_port()']]],
+  ['get_5fread_5ffd',['get_read_fd',['../classtacopie_1_1self__pipe.html#a301b416f5f8236a383065b381385b88c',1,'tacopie::self_pipe']]],
+  ['get_5fsocket',['get_socket',['../classtacopie_1_1tcp__client.html#a1a3834deb1d263ec5816066f74286298',1,'tacopie::tcp_client::get_socket(void)'],['../classtacopie_1_1tcp__client.html#a9cf1f3ccf43f9a0a883a17b15e3668d6',1,'tacopie::tcp_client::get_socket(void) const'],['../classtacopie_1_1tcp__server.html#a39a51b9203d42babfd9c4c1a0f4cc340',1,'tacopie::tcp_server::get_socket(void)'],['../classtacopie_1_1tcp__server.html#a373aec294e24a52c3ef6c44920af36e2',1,'tacopie::tcp_server::get_socket(void) const']]],
+  ['get_5ftype',['get_type',['../classtacopie_1_1tcp__socket.html#a4f663be51b845520505bc20a88b411ee',1,'tacopie::tcp_socket']]],
+  ['get_5fwrite_5ffd',['get_write_fd',['../classtacopie_1_1self__pipe.html#ab36a4deb45bb408988f26315aedc0d74',1,'tacopie::self_pipe']]]
+];